rename notifications manager file to match component name
