+++ src/frontend/src/components/NotificationManager/components/CreateNSTreeSuccessNotif/index.tsx
@@ -1,7 +1,12 @@
 import NextLink from "next/link";
+import { ROUTES } from "src/common/routes";
 import { StyledButton } from "./style";
 
-export const CreateNSTreeSuccessNotif = (): JSX.Element => (
+interface Props {
+  onDismiss(): void;
+}
+
+export const CreateNSTreeSuccessNotif = ({ onDismiss }: Props): JSX.Element => (
   <>
     <span>
       Your tree is being created. It may take up to 12 hours to process. To
@@ -12,6 +17,7 @@ export const CreateNSTreeSuccessNotif = (): JSX.Element => (
         <StyledButton
           sdsType="primary"
           sdsStyle="minimal"
+          onClick={onDismiss}
         >
           VIEW MY TREES
         </StyledButton>
