+++ src/frontend/src/components/NotificationManager/components/CreateNSTreeSuccessNotif/style.ts
